Fix PetGoalLookAtPlayer not working for 1.14-1.16

+++ modules/v1_13_R2/src/com/dsh105/echopet/compat/nms/v1_13_R2/entity/ai/PetGoalLookAtPlayer.java
@@ -90,8 +90,7 @@ public class PetGoalLookAtPlayer extends APetGoalLookAtPlayer{
 	
 	@Override
 	public void tick(){
-		// 1.9: this.a.getControllerLook().a(this.b.locX, this.b.locY + this.b.getHeadHeight(), this.b.locZ, this.a.cE(), this.a.N());
-		this.pet.getControllerLook().a(this.player.locX, this.player.locY + (double) this.player.getHeadHeight(), this.player.locZ, 10.0F, (float) this.pet.K());
+		this.pet.getControllerLook().a(this.player.locX, this.player.locY + this.player.getHeadHeight(), this.player.locZ, 10.0F, (float) this.pet.K());
 		--this.ticksLeft;
 	}
 }
